Refactor ConfiguresPaymentGateway trait tests
