+++ tests/Unit/Config/PaymentMethodRequestsTraitTest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Payavel\Checkout\Tests\Unit\Config;
+
+use Payavel\Checkout\Tests\Unit\TestPaymentMethodRequestsTrait;
+
+class PaymentMethodRequestsTraitTest extends TestPaymentMethodRequestsTrait
+{
+    protected $driver = 'config';
+}
