+++ tests/Unit/Config/WalletRequestsTraitTest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Payavel\Checkout\Tests\Unit\Config;
+
+use Payavel\Checkout\Tests\Unit\TestWalletRequestsTrait;
+
+class WalletRequestsTraitTest extends TestWalletRequestsTrait
+{
+    protected $driver = 'config';
+}
