+++ tests/Unit/Database/PaymentMethodRequestsTraitTest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Payavel\Checkout\Tests\Unit\Database;
+
+use Payavel\Checkout\Tests\Unit\TestPaymentMethodRequestsTrait;
+
+class PaymentMethodRequestsTraitTest extends TestPaymentMethodRequestsTrait
+{
+    protected $driver = 'database';
+}
